Crate sunrise_libkern

Types shared by user and kernel

Modules

error
Kernel errors

nr
Syscall numbers

process
Data-structures related to process syscalls.

Structs

MemoryAttributes
Low-level attributes of a memory mapping.

MemoryInfo
The structure returned by the query_memory syscall.

MemoryPermissions
Memory permissions of a memory area.

MemoryState
Represents the current state of a memory region: why is it allocated, and what operations are allowed.

SYSCALL_NAMES
A table associating a syscall name string for every syscall number.

TLS
Thread Local Storage region.

Enums

MemoryType
The type of this memory area.

Type Definitions

IpcBuffer
Buffer used for Inter Process Communication. Kernel reads, interprets, and copies data from/to it.