Module sunrise_libkern::process

Data-structures related to process syscalls.

Structs

KipHeader
Header for Kernel Builtins. Can be found in the .kip_header section of our ELFs. Nintendo KIPs start with a (slightly different, but functionally equivalent) header.

ProcInfo
Informations necessary for the create_process syscall.

ProcInfoFlags
Miscelaneous flags.

ProcessCategory
Category of the process.

ProcessInfoType
Kind of information to extract from a process wit get_process_info.

ProcessState
The state the process is currently in.

Enums

PoolPartition
Kernel memory pool.

ProcInfoAddrSpace
Address space type to use when creating a process.